+++ esp32/components/rpg/rpg/Coordinates.h
@@ -11,7 +11,7 @@ namespace rpg
 class Coordinates
 {
   public:
-    void change_tile_by(int dx, int dy)
+    void change_tile_by(int16_t dx, int16_t dy)
     {
         value_tile_x += dx;
         value_tile_y += dy;
@@ -20,7 +20,7 @@ class Coordinates
         value_y += convert_to_xy(dy);
     }
 
-    void change_xy_by(int dx, int dy)
+    void change_xy_by(int16_t dx, int16_t dy)
     {
         value_x += dx;
         value_y += dy;
@@ -29,22 +29,22 @@ class Coordinates
         value_tile_y = convert_to_tile(value_y);
     }
 
-    int tile_x() const
+    int16_t tile_x() const
     {
         return value_tile_x;
     }
 
-    int tile_y() const
+    int16_t tile_y() const
     {
         return value_tile_y;
     }
 
-    int x() const
+    int16_t x() const
     {
         return value_x;
     }
 
-    int y() const
+    int16_t y() const
     {
         return value_y;
     }
@@ -52,7 +52,7 @@ class Coordinates
     /**
      * Check that coordinates lay within the specified box using tile units.
      */
-    bool within_tile(int min_x, int min_y, int max_x, int max_y)
+    bool within_tile(int16_t min_x, int16_t min_y, int16_t max_x, int16_t max_y)
     {
         if (value_tile_x < min_x)
             return false;
@@ -72,7 +72,7 @@ class Coordinates
     /**
      * Check that coordinates lay within the specified box using pixels.
      */
-    bool within_xy(int min_x, int min_y, int max_x, int max_y)
+    bool within_xy(int16_t min_x, int16_t min_y, int16_t max_x, int16_t max_y)
     {
         if (value_x < min_x)
             return false;
@@ -90,27 +90,27 @@ class Coordinates
     }
 
   protected:
-    Coordinates(int x, int y, int tile_x, int tile_y)
+    Coordinates(int16_t x, int16_t y, int16_t tile_x, int16_t tile_y)
         : value_tile_x{tile_x}, value_tile_y{tile_y}, value_x{x}, value_y{y}
     {
     }
 
-    static int convert_to_tile(int xy)
+    static int16_t convert_to_tile(int16_t xy)
     {
         return xy / DISPLAY_TILE_WIDTH -
                ((xy % DISPLAY_TILE_WIDTH) < 0 ? 1 : 0);
     }
 
-    static int convert_to_xy(int tile)
+    static int16_t convert_to_xy(int16_t tile)
     {
         return tile * DISPLAY_TILE_WIDTH;
     }
 
   private:
-    int value_tile_x;
-    int value_tile_y;
-    int value_x;
-    int value_y;
+    int16_t value_tile_x;
+    int16_t value_tile_y;
+    int16_t value_x;
+    int16_t value_y;
 };
 
 /**
@@ -121,13 +121,13 @@ class GlobalCoordinates : public Coordinates
   public:
     using Coordinates::Coordinates;
 
-    static GlobalCoordinates tile(int tile_x, int tile_y)
+    static GlobalCoordinates tile(int16_t tile_x, int16_t tile_y)
     {
         return GlobalCoordinates(convert_to_xy(tile_x), convert_to_xy(tile_y),
                                  tile_x, tile_y);
     }
 
-    static GlobalCoordinates xy(int x, int y)
+    static GlobalCoordinates xy(int16_t x, int16_t y)
     {
         return GlobalCoordinates(x, y, convert_to_tile(x), convert_to_tile(y));
     }
@@ -146,13 +146,13 @@ class LocalCoordinates : public Coordinates
   public:
     using Coordinates::Coordinates;
 
-    static LocalCoordinates tile(int tile_x, int tile_y)
+    static LocalCoordinates tile(int16_t tile_x, int16_t tile_y)
     {
         return LocalCoordinates(convert_to_xy(tile_x), convert_to_xy(tile_y),
                                 tile_x, tile_y);
     }
 
-    static LocalCoordinates xy(int x, int y)
+    static LocalCoordinates xy(int16_t x, int16_t y)
     {
         return LocalCoordinates(x, y, convert_to_tile(x), convert_to_tile(y));
     }
@@ -171,13 +171,13 @@ class ScreenCoordinates : public Coordinates
   public:
     using Coordinates::Coordinates;
 
-    static ScreenCoordinates tile(int tile_x, int tile_y)
+    static ScreenCoordinates tile(int16_t tile_x, int16_t tile_y)
     {
         return ScreenCoordinates(convert_to_xy(tile_x), convert_to_xy(tile_y),
                                  tile_x, tile_y);
     }
 
-    static ScreenCoordinates xy(int x, int y)
+    static ScreenCoordinates xy(int16_t x, int16_t y)
     {
         return ScreenCoordinates(x, y, convert_to_tile(x), convert_to_tile(y));
     }
